requirements.txt
Django==5.1.2
gunicorn
whitenoise
dj-database-url
python-dotenv
psycopg2-binary
crispy-bootstrap5==2024.10
django-crispy-forms==2.3
pillow==11.0.0
django-import-export==4.1.1
python-decouple==3.8
openai==25.0.1